requirements.txt
streamlit
pillow
requests
joblib
scikit-learn==1.2.2